### 2025-06-17 #186 @editorialbot
[REVIEW]: SequentialSamplingModels.jl: Simulating and Evaluating Cognitive Models of Response Times in Julia
**Submitting author:** @kiante-fernandez (<a href="http://orcid.org/0000-0002-8493-880X">Kianté Fernandez</a>)
**Repository:** https://github.com/kiante-fernandez/SSM-paper
**Branch with paper.md** (empty if default branch): main
**Version:** v2
