### 2025-06-17 16fabfd
re-test paper render for review
