+++ paper/bib.tex
@@ -1,5 +1,4 @@
 % **************GENERATED FILE, DO NOT EDIT**************
 
 \bibliographystyle{juliacon}
-
 \bibliography{ref.bib}
