+++ paper/header.tex
@@ -5,16 +5,16 @@
 \author[1]{Kianté Fernandez}
 \author[2]{Dominique Makowski}
 \author[3]{Christopher Fisher}
-\affil[1]{University of California, Los Angeles}
+\affil[1]{Department of Psychology, University of California, Los Angeles}
 \affil[2]{School of Psychology, University of Sussex, Brighton, UK}
 \affil[3]{Independent Researcher}
 
-\keywords{cognitive models,response time, sequential sampling models, evidence accumulation, decision-making}
+\keywords{cognitive models, response time, sequential sampling models, evidence accumulation, decision-making}
 
 \hypersetup{
-pdftitle = {My JuliaCon proceeding},
+pdftitle = {SequentialSamplingModels.jl: Simulating and Evaluating Cognitive Models of Response Times in Julia},
 pdfsubject = {JuliaCon 2024 Proceedings},
-pdfauthor = {1st author, 2nd author, 3rd author},
-pdfkeywords = {cognitive models,response time, sequential sampling models, evidence accumulation, decision-making},
+pdfauthor = {Kianté Fernandez, Dominique Makowski, Christopher Fisher},
+pdfkeywords = {cognitive models, response time, sequential sampling models, evidence accumulation, decision-making},
 }
 
