+++ paper/paper.yml
@@ -21,7 +21,7 @@ affiliations:
     index: 1
   - name: School of Psychology, University of Sussex, Brighton, UK
     index: 2
-  - name: Parallax Advanced Research
+  - name: Independent Researcher
     index: 3
 date: 7 November 2024
 bibliography: ref.bib
