### 2025-06-17 bef3ea1
add paper.md
