### 2025-06-17 #186 @kiante-fernandez
> Hi [@kiante-fernandez](https://github.com/kiante-fernandez). Can you change the repository/paper branch so that the `editorialbot` can correctly locate it? Thanks!

@fcdimitr I think you can only change that. The paper is in the https://github.com/kiante-fernandez/SSM-paper repo
