### 2025-12-06 #186 @rsenne
@itsdfish many citations do not have DOIs please add them where possible. Here are a few I found:

10.1162/neco.2008.12-06-420 (Ratcliff & Mckoon 2008)|
10.1037/0033-295X.85.2.59 (Ratcliff 1978)
10.48550/arXiv.2308.09769 (Pidgeons.jl)
10.32614/RJ-2014-005 (rwiener package)

This is the last remaining issue after updating my checklist.
