### 2025-12-07 #186 @itsdfish
@rsenne thank you for pointing that out. @kiante-fernandez, added the DIOs. For a few references, I also changed the field journaltitle to journal so that the journal names properly display. I think we should be good to go.
