### 2025-12-07 #186 @rsenne
@itsdfish @kiante-fernandez where were these changes made? I can't seem to generate the pdf to verify and the [paper repo](https://github.com/kiante-fernandez/SSM-paper) does not show a recent update?
