### 2025-12-08 dcf3646
Refactor paper structure and update references: - Updated LaTeX file to improve citation formatting and enhance clarity in the introduction and examples. - Changed author affiliation from "Independent Researcher" to "Parallax Advanced Research." - Revised references in the bibliography for consistency and accuracy, including updates to DOIs and URLs. - Enhanced example descriptions and added parameter explanations in the code examples for better understanding.
