+++ paper/paper.md
@@ -1,204 +1,0 @@
-# Introduction
-
-Sequential sampling models (SSMs) are widely used in cognitive science
-due to their ability to describe dissociable processes underlying a wide
-variety of capacities, including memory retrieval, visual perception,
-and decision making ([@forstmann2016sequential]). These models typically
-describe decision-making as a stochastic, dynamic evidence accumulation
-process which evolves until evidence for one option reaches an evidence
-threshold
-([@smith2000stochastic; @Ratcliff2008TheDD; @Ratcliff1978ATO]). By doing
-so, SSMs provide a generative model for response time (RT) distributions
-of actions made by organisms capturing both the speed of the response
-and the response itself. Figure 1 illustrates the latent evidence
-accumulation process for a hypothetical decision between a Margherita
-pizza and a pineapple pizza. Evidence accumulates stochastically (i.e.,
-randomly) until it reaches either the upper boundary, triggering the
-selection of the Margherita, or the lower boundary, triggering the
-selection of the pineapple pizza. While these models hold many promising
-applications, their implementation and use remain a technical challenge.
-
-The widespread interest and continuous use of SSMs across cognitive
-science have spurred the development of various software packages that
-lower the barrier for simulating, estimating, and comparing existing
-SSMs. These software implementations range widely, from those designed
-to provide maximum flexibility to end-users ([@Shinn2020AFF]) to those
-that streamline the entire analysis of SSMs ([@Fengler2022BeyondDD]).
-Toolboxes have also been developed across multiple programming languages
-commonly used for scientific computing, including MATLAB
-([@vandekerckhove2008diffusion]), Python
-([@wiecki2013hddm; @Shinn2020AFF; @Fengler2022BeyondDD; @murrow2024pybeam]),
-and R
-([@wabersich2014rwiener; @stevenson2024emc2; @hBayesDM; @singmann2018package; @hartmann2021partial]).
-
-![An example of applying the sequential sampling modeling framework to a
-choice between two pizzas. The decision-maker samples evidence in favor
-of both options until reaching a decision
-boundary.](figures/SSM_example_fig1.png){#fig:exampleSSM width="100%"}
-
-# SequentialSamplingModels.jl
-
-While existing methods can already simulate a wide range of SSMs across
-multiple languages, developments in Julia have either not been
-established or lack maintenance. SequentialSamplingModels (SSM.jl)
-provides the first unified interface for simulating validated SSMs,
-drawn from the literature to represent those most widely used. The
-package offers a user-friendly method for simulating a range of SSMs of
-interest to researchers. It is the first attempt to integrate a package
-into the Julia ecosystem for simulating and evaluating a popular model
-of decision-making.
-
-# Availability, Development, and Documentation
-
-SequentialSamplingModels.jl is available on the Julia package registry.
-The package is maintained on GitHub at
-<https://github.com/itsdfish/SequentialSamplingModels.jl>, and the
-documentation can be found at
-<https://itsdfish.github.io/SequentialSamplingModels.jl/dev/>. For a
-more comprehensive exploration of the package functionalities and
-further details on future developments, the user is invited to consult
-the package README, documentation, and issues.
-
-# Example one: simulating choices and response times
-
-SequentialSamplingModels.jl is an extension of the established
-functionalities within the Distributions package
-[@dahua_lin_2024_13838344], designed for generating multivariate
[…]
