+++ paper/paper.tex
@@ -2,8 +2,8 @@
 % JuliaCon proceedings template
 \documentclass{juliacon}
 \usepackage{caption}
-\usepackage[numbers]{natbib}  % Add citation management
-\bibliographystyle{juliacon}
+\usepackage[round]{natbib}  % Add citation management
+\bibliographystyle{plainnat} % Format as Last name, First name
 
 \setcounter{page}{1}
 
@@ -14,26 +14,25 @@
 \maketitle
 
 \begin{abstract}
-
-Sequential sampling models (SSMs) are a widely used framework describing decision-making as a stochastic, dynamic process of evidence accumulation. SSMs popularity across cognitive science has driven the development of various software packages that lower the barrier for simulating, estimating, and comparing existing SSMs. Here, we present a software tool, SequentialSamplingModels.jl (SSM.jl), designed to make SSM simulations more accessible to Julia users, and to integrate with the Julia ecosystem. We demonstrate the basic use of SSM.jl for simulation, plotting, and Bayesian inference.
+Sequential sampling models (SSMs) are a widely used framework describing decision-making as a stochastic, dynamic process of evidence accumulation. SSMs' popularity across cognitive science has driven the development of various software packages that lower the barrier for simulating, estimating, and comparing existing SSMs. Here, we present a software tool, SequentialSamplingModels.jl (SSM.jl), designed to make SSM simulations more accessible to Julia users, and to integrate with the Julia ecosystem. We demonstrate the basic use of SSM.jl for simulation, plotting, and Bayesian inference.
 
 \end{abstract}
 
 \section{Introduction}
-Sequential sampling models (SSMs) are widely used in cognitive science due to their ability to describe dissociable processes underlying a wide variety of capacities, including memory retrieval, visual perception, and decision making (\cite{forstmann2016sequential}). These models typically describe decision-making as a stochastic, dynamic evidence accumulation process which evolves until evidence for one option reaches an evidence threshold (\cite{smith2000stochastic,Ratcliff2008TheDD,Ratcliff1978ATO}). By doing so, SSMs provide a generative model for response time (RT) distributions of actions made by organisms capturing both the speed of the response and the response itself. Figure 1 illustrates the latent evidence accumulation process for a hypothetical decision between a Margherita pizza and a pineapple pizza. Evidence accumulates stochastically (i.e., randomly) until it reaches either the upper boundary, triggering the selection of the Margherita, or the lower boundary, triggering the selection of the pineapple pizza. While these models hold many promising applications, their implementation and use remain a technical challenge.
[…]
