+++ paper/bib.tex
@@ -1,4 +1,5 @@
 % **************GENERATED FILE, DO NOT EDIT**************
 
 \bibliographystyle{juliacon}
+
 \bibliography{ref.bib}
