### 2025-12-10 #186 @fcdimitr
Thank you all!
@kiante-fernandez there are two things before acceptance:
1. Could you check the [error that appears](https://github.com/JuliaCon/proceedings-papers/actions/runs/20113211176/job/57715947880) when compiling the PDF? There seems to be an issue with a bibtex entry.
2. Could you please check the Zenodo archive that is required for acceptance?
You need to follow [the instructions here](https://juliacon.github.io/proceedings-guide/author/#publication_phase) so that there is a permanent DOI for the final (accepted) version of the code & paper.

Then, we can accept it and close this issue. Thanks!
