### 2025-12-10 984accb
Fix formatting issues in references and update DOIs in ref.bib
