+++ paper/ref.bib
@@ -131,8 +131,7 @@ publisher = {eLife Sciences Publications, Ltd},
   journal={eLife},
   year={2020},
   volume={9},
-doi={https://doi.org/10.7554/eLife.56938
-}
+  doi={https://doi.org/10.7554/eLife.56938}
 }
 
 @article{Fengler2022BeyondDD,
@@ -261,7 +260,6 @@ doi={https://doi.org/10.3758/BRM.40.1.61}
 
 
 @article{PlotsJL,
-  doi = {https://doi.org/10.5334/jors.431},
   url = {https://openresearchsoftware.metajnl.com/articles/10.5334/jors.431/},
   author = {Christ, Simon and Schwabeneder, Daniel and Rackauckas, Christopher and Borregaard, Michael Krabbe and Breloff, Thomas},
   keywords = {Graphics (cs.GR), FOS: Computer and information sciences, FOS: Computer and information sciences, I.3.3},
@@ -269,9 +267,7 @@ doi={https://doi.org/10.3758/BRM.40.1.61}
   publisher = {Journal of Open Research Software},
   year = {2023},
   copyright = {Creative Commons Attribution 4.0 International},
-doi={
-https://doi.org/10.48550/arXiv.2204.08775
-}
+  doi = {https://doi.org/10.5334/jors.431}
 }
 
 @article{singmann2018package,
@@ -303,8 +299,7 @@ https://doi.org/10.48550/arXiv.2204.08775
   pages={408--463},
   year={2000},
   publisher={Elsevier},
-doi={
-https://doi.org/10.1006/jmps.1999.1260}
+  doi={https://doi.org/10.1006/jmps.1999.1260}
 }
 
 @article{murrow2024pybeam,
@@ -316,11 +311,10 @@ https://doi.org/10.1006/jmps.1999.1260}
   pages={2636--2656},
   year={2024},
   publisher={Springer},
-doi={
-https://doi.org/10.3758/s13428-023-02162-w}
-}
+  doi={https://doi.org/10.3758/s13428-023-02162-w}
+  }
 
-@article{wabersich2014rwiener=,
+@article{wabersich2014rwiener,
 	title = {The {RWiener} Package: an R Package Providing Distribution Functions for the Wiener Diffusion Model},
 	volume = {6},
 	issn = {2073-4859},
