### 2025-12-10 #186 @kiante-fernandez
@fcdimitr I have addressed the bibtex bug and push those changes.

Also here is Zenodo meeting those requirements: https://doi.org/10.5281/zenodo.14070013
