### 2025-12-11 #186 @fcdimitr
Thank you, @kiante-fernandez. The [remaining issue](https://github.com/JuliaCon/proceedings-papers/actions/runs/20118238052/job/57732640259) appears to be related to figure paths: the workflow log is reporting errors with the relative paths to the images. Please try running the `generate pdf` command here while debugging, and iterate until the manuscript compiles successfully.

Regarding the Zenodo DOI, it appears to point to an older version of the code. Since the package has evolved following the review and [the issues raised in the repository](https://github.com/itsdfish/SequentialSamplingModels.jl/issues/126#issuecomment-3116798664), please issue an updated verion/DOI on Zenodo, corresponding to the version that incorporates those changes (the one the reviewers evaluated).

If anything is unclear or you run into further issues, feel free to reach out. Thank you!
