### 2025-12-11 #186 @rsenne
@kiante-fernandez Pretty sure that based on the error message:

`\includegraphics[width=1\linewidth]{paper//figures/SSM_example_fig1.png}\includegraphics[width=1\linewidth]{paper//figures/SSM_example_fig1.png}`

should be

`\includegraphics[width=1\linewidth]{figures/SSM_example_fig1.png}`
