### 2025-12-11 a4ef74a
Fix image path in figure inclusion for SSM example
