+++ paper/paper.tex
@@ -25,7 +25,7 @@ The widespread interest and continuous use of SSMs across cognitive science have
 
 \begin{figure}
     \centering
-    \includegraphics[width=1\linewidth]{paper//figures/SSM_example_fig1.png}
+    \includegraphics[width=1\linewidth]{paper/figures/SSM_example_fig1.png}
     \caption{An example of applying the sequential sampling modeling framework to a choice between two pizzas. The decision-maker samples evidence in favor of both options until reaching a decision boundary.}
     \label{fig:exampleSSM}
 \end{figure}
