### 2025-12-11 112b993
Fix figure paths in paper.tex to ensure correct image rendering
