### 2025-12-15 #186 @fcdimitr
Thank you all, I think we are almost there! Could you please fix the DOI issues above in the references?
