### 2025-12-15 3ced376
Fix DOI formatting in references to remove URLs in ref.bib
