+++ paper/ref.bib
@@ -131,7 +131,7 @@ publisher = {eLife Sciences Publications, Ltd},
   journal={eLife},
   year={2020},
   volume={9},
-  doi={https://doi.org/10.7554/eLife.56938}
+  doi={10.7554/eLife.56938}
 }
 
 @article{Fengler2022BeyondDD,
@@ -161,7 +161,7 @@ publisher = {eLife Sciences Publications, Ltd},
   pages={911--936},
   year={2020},
   publisher={Springer},
-doi={https://doi.org/10.3758/s13423-020-01719-6}
+doi={10.3758/s13423-020-01719-6}
 }
 
 @article{ratcliff2016diffusion,
@@ -190,7 +190,7 @@ doi={https://doi.org/10.3758/s13423-020-01719-6}
   pages={153--178},
   year={2008},
   publisher={Elsevier},
-doi={https://doi.org/10.1016/j.cogpsych.2007.12.002}
+doi={10.1016/j.cogpsych.2007.12.002}
 }
 @article{wiecki2013hddm,
   title={HDDM: Hierarchical Bayesian estimation of the drift-diffusion model in Python},
@@ -200,7 +200,7 @@ doi={https://doi.org/10.1016/j.cogpsych.2007.12.002}
   pages={55610},
   year={2013},
   publisher={Frontiers},
-doi={https://doi.org/10.3389/fninf.2013.00014}
+doi={10.3389/fninf.2013.00014}
 }
 @article{vandekerckhove2008diffusion,
   title={Diffusion model analysis with MATLAB: A DMAT primer},
@@ -211,7 +211,7 @@ doi={https://doi.org/10.3389/fninf.2013.00014}
   pages={61--72},
   year={2008},
   publisher={Springer},
-doi={https://doi.org/10.3758/BRM.40.1.61}
+doi={10.3758/BRM.40.1.61}
 }
 
 @article{stevenson2024emc2,
@@ -232,7 +232,7 @@ doi={https://doi.org/10.3758/BRM.40.1.61}
   pages={641--666},
   year={2016},
   publisher={Annual Reviews},
-  doi={https://doi.org/10.1146/annurev-psych-122414-033645}
+  doi={10.1146/annurev-psych-122414-033645}
 }
 
 @article{hBayesDM,
@@ -267,7 +267,7 @@ doi={https://doi.org/10.3758/BRM.40.1.61}
   publisher = {Journal of Open Research Software},
   year = {2023},
   copyright = {Creative Commons Attribution 4.0 International},
-  doi = {https://doi.org/10.5334/jors.431}
+  doi = {10.5334/jors.431}
 }
 
 @article{singmann2018package,
@@ -277,6 +277,7 @@ doi={https://doi.org/10.3758/BRM.40.1.61}
 	journal = {R package version 0.4-9. {URL} http://{CRAN}. R-project. org/package= rtdists},
 	author = {Singmann, Henrik and Brown, S. and Gretton, M. and Heathcote, A. and Voss, A. and Voss, J. and Terry, A.},
 	date = {2016},
+	doi = {10.32614/cran.package.rtdists},
 }
 
 @article{hartmann2021partial,
@@ -287,7 +288,7 @@ doi={https://doi.org/10.3758/BRM.40.1.61}
   pages={102550},
   year={2021},
   publisher={Elsevier},
-  doi={https://doi.org/10.1016/j.jmp.2021.102550}
+  doi={10.1016/j.jmp.2021.102550}
 }
 
 @article{smith2000stochastic,
@@ -299,7 +300,7 @@ doi={https://doi.org/10.3758/BRM.40.1.61}
   pages={408--463},
   year={2000},
[…]
