### 2025-12-15 39f7c25
Add two more missing DOIs to references in ref.bib
