+++ paper/ref.bib
@@ -22,7 +22,8 @@
   volume={15},
   number={1},
   pages={1593--1623},
-  year={2014}
+  year={2014},
+  doi={10.5555/2627435.2638586}
 }
 
 @article{Ratcliff1978ATO,
@@ -123,6 +124,7 @@ publisher = {eLife Sciences Publications, Ltd},
   year      = {2018},
   url       = {http://proceedings.mlr.press/v84/ge18b.html},
   biburl    = {https://dblp.org/rec/bib/conf/aistats/GeXG18},
+  doi       = {10.17863/CAM.42246},
 }
 
 @article{Shinn2020AFF,
