### 2025-12-15 1ad2ac3
Fix DOI entry by replacing doi field with url in ref.bib
