+++ paper/ref.bib
@@ -23,7 +23,7 @@
   number={1},
   pages={1593--1623},
   year={2014},
-  doi={10.5555/2627435.2638586}
+  url={https://dl.acm.org/doi/10.5555/2627435.2638586}
 }
 
 @article{Ratcliff1978ATO,
