### 2025-12-15 #186 @editorialbot
:wave: @JuliaCon/jcon-eics, this paper is ready to be accepted and published.
